STENGLOC-2177 Reviewer: None implement system and process cmds

+++ tasklets/agent/cmds/system/tasklets/reboot.py
@@ -6,9 +6,13 @@ def match(q, i, params, tags):
     return params['subcmd'] == 'reboot'
 
 def main(q, i, params, tags):
-    q.logger.log("system reboot params:%s tags:%s"%(params,tags))
-    params["returnmessage"] = 'Successfully executed command system reboot' 
-    params["returncode"] = 0
+    q.logger.log("system reboot params:%s tags:%s"%(params,tags), tags = 'tasknr:%s'%params['tasknr'])
+    
+    cmd = 'shutdown /r' if q.platform.isWindows() else 'reboot'
+    output = q.system.process.executeAsync(cmd, argsInCommand = True)
+    output.wait()
+    params["returnmessage"] = 'Successfully executed command system reboot' if output.returncode == 0 else ''.join(output.stderr.readlines())
+    params["returncode"] = output.returncode
